Replace all use of UnityEngine.Assertions with exceptions
The goal of this assertions library is to have validation logic that gets conditionally compiled out in release builds, on the assumption that those validation checks are not necessary in release builds b/c they've already been tested in debug builds. However, these performance savings are generally minimal, and they come at the cost of _breaking_ that assumption and not getting any useful error message about it. Assertions should only be used in unit tests, failed runtime assertions should always throw exceptions.

+++ src/UnityUtil/Legal/SingleDialogConsentManager.cs
@@ -7,7 +7,6 @@ using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Threading.Tasks;
 using UnityEngine;
-using UnityEngine.Assertions;
 using UnityEngine.Events;
 using UnityUtil.DependencyInjection;
 using UnityUtil.Logging;
@@ -192,7 +191,8 @@ public class SingleDialogConsentManager : MonoBehaviour, IConsentManager
     [Button, Conditional("UNITY_EDITOR")]
     public void ClearConsentPreferences()
     {
-        Assert.IsTrue(Application.isPlaying, "Consent can only be cleared in Play Mode, so that initializable dependencies are registered.");
+        if (!Application.isPlaying)
+            throw new InvalidOperationException("Consent can only be cleared in Play Mode, so that initializable dependencies are registered.");
 
         foreach (IInitializableWithConsent initializableWithConsent in _initializablesWithConsent!)
             _localPreferences!.DeleteKey(initializableWithConsent.ConsentPreferenceKey);
